- ADD: Games can now be sold - UPDATE: Project Settings now contains values for selling Games; - Update: IncomeManager now calculates Income for game on the market;

+++ Assets/Scripts/saveManager.cs
@@ -499,7 +499,7 @@ public class saveManager : MonoBehaviour
             float _companyMoney = saveData.GetFloat("CompanyMoney");
 
             _companyValuesManager.companyReputaion = _companyReputation;
-            _companyValuesManager.companyCash = _companyMoney;
+            _companyValuesManager.companyCash = (decimal)_companyMoney;
 
             //Load & Instantiate Player
             int playerPrefabIndex = saveData.GetInt("Player CharacterPrefab_");
